switch html - rte
