+++ apps/web/app/composables/useHtmlEditorMode/types.ts
@@ -8,3 +8,5 @@ export type UseHtmlEditorModeOptions = {
 export type HtmlToken =
   | { kind: 'comment' | 'doctype' | 'processing-instruction' }
   | { kind: 'tag'; tagName: string; attributesSource: string; isClosing: boolean; isSelfClosing: boolean };
+
+export type QuoteState = 'none' | 'single' | 'double';
